requirements.txt
torch~=1.11.0
PyYAML~=6.0
wandb~=0.12.16
utils~=1.0.1
tqdm~=4.62.3
opencv-python~=4.5.5.64
matplotlib~=3.5.1
numpy~=1.22.3
pandas~=1.4.0
seaborn~=0.11.2
Pillow~=9.1.0
torchvision~=0.12.0
requests~=2.27.1
thop~=0.0.31.post2005241907
twilio~=7.9.0